+++ test/OpenTelemetry.Tests/Trace/SamplersTest.cs
@@ -42,7 +42,7 @@ namespace OpenTelemetry.Trace.Tests
             var link = new ActivityLink(parentContext);
 
             Assert.Equal(
-                SamplingDecision.RecordAndSampled,
+                SamplingDecision.RecordAndSample,
                 new AlwaysOnSampler().ShouldSample(new SamplingParameters(parentContext, this.traceId, "Another name", ActivityKindServer, null, new List<ActivityLink> { link })).Decision);
         }
 
@@ -61,7 +61,7 @@ namespace OpenTelemetry.Trace.Tests
             var link = new ActivityLink(parentContext);
 
             Assert.Equal(
-                SamplingDecision.NotRecord,
+                SamplingDecision.Drop,
                 new AlwaysOffSampler().ShouldSample(new SamplingParameters(parentContext, this.traceId, "Another name", ActivityKindServer, null, new List<ActivityLink> { link })).Decision);
         }